Removed slider.js from modules. Moved function to index.js.
